refactor(volPyrolysis): split the solid flux limiter into its own collaborator
The two-sided solid flux limiter arrived whole with the bounded-transport
work and is self-contained, so it moves out of volPyrolysis now, before
anything comes to depend on its current shape. volPyrolysis was already
the largest class in the tree; this takes ~740 lines off it.

SolidFluxLimiter owns solidVolFlux(), accumulateFaceFlux(),
solidFluxBudgets(), solidDonorLimit(), solidReceiverLimit(),
applySolidFaceLimit() and reportSolidFluxLimiter(), and the orchestrator
that sweeps them becomes its public limit(). The bodies are unchanged;
solidFluxBudgets() takes the flux it reads as a parameter now that it is
no longer a member of the class that holds it. Plain composition, as with
STmodel_/HTmodel_/solidChemistry_ - there is one limiter algorithm, so no
runtime-selection table.

advectSolidFields, nSolidFluxLimiterCorrectors and minPorosity move with
it and are read from the same pyrolysisCoeffs keys, so the dictionary is
unchanged. Their banner lines move too, which groups the three settings
the limiter owns instead of scattering them; the wording is untouched.
solidStateTolerance is read on both sides - volPyrolysis's fail-fast
guards need it as well, and only volPyrolysis logs it.

phiSolid_ stays in volPyrolysis: solveSpeciesMass() and preSolveEnergy()
read it outside the limiter's scope, so limit() returns the flux and the
caller stores it.

The fail-fast guards move to solidStateGuards.H as free functions, taking
the mesh and time that were implicit in the member versions. Six call
sites across recoverPorosity(), solveSpeciesMass(), preSolveEnergy() and
postSolveEnergy() are updated. The fatal message body is unchanged;
FatalErrorInFunction now names the free function.

No intended behaviour change. convergingJam to endTime 0.5 (250 steps)
was run against this tip and against the pre-refactor tip: every written
time directory and postProcessing output is bit-identical, and the solver
logs differ only in the position of the relocated advectSolidFields
banner line - 240/250 limiter-active steps, 250 consistency lines, no
"NOT converged", no fatal, on both.

Refs #88
