+++ porousGasificationMedia/pyrolysisModels/Make/files
@@ -5,5 +5,6 @@ BCs/fixedYm/fixedYmFvPatchScalarField.C
 pyrolysisModel/heterogeneousPyrolysisModel/heterogeneousPyrolysisModel.C
 pyrolysisModel/heterogeneousPyrolysisModel/heterogeneousPyrolysisModelNew.C
 pyrolysisModel/volPyrolysis/volPyrolysis.C
+pyrolysisModel/volPyrolysis/solidFluxLimiter/solidFluxLimiter.C
 
 LIB = $(FOAM_USER_LIBBIN)/libHGSheterogeneousPyrolysisModels
