+++ porousGasificationMedia/pyrolysisModels/pyrolysisModel/volPyrolysis/volPyrolysis.C
@@ -39,8 +39,8 @@ License
 #include "processorPolyPatch.H"
 #include "processorCyclicPolyPatch.H"
 #include "upwind.H"
-#include "slicedSurfaceFields.H"
-#include "syncTools.H"
+
+#include "solidStateGuards.H"
 
 #include "BCs/fixedSolidH/fixedSolidHFvPatchScalarField.H"
 #include "BCs/fixedYm/fixedYmFvPatchScalarField.H"
@@ -133,583 +133,9 @@ void volPyrolysis::deriveYiFromYm()
     }
 }
 
-tmp<surfaceScalarField> volPyrolysis::solidVolFlux() const
-{
-    tmp<surfaceScalarField> tSolidVolFlux
-    (
-        new surfaceScalarField
-        (
-            IOobject
-            (
-                "solidVolFlux",
-                time_.timeName(),
-                mesh_,
-                IOobject::NO_READ,
-                IOobject::NO_WRITE
-            ),
-            mesh_,
-            dimensionedScalar("zero", dimVolume/dimTime, 0.0)
-        )
-    );
-
-    if (advectSolidFields_)
-    {
-        tSolidVolFlux.ref() = mesh_.Sf() & fvc::interpolate(Us_, "Us");
-    }
-
-    return tSolidVolFlux;
-}
-
-void volPyrolysis::accumulateFaceFlux
-(
-    const surfaceScalarField& phi,
-    const surfaceScalarField& lambda,
-    scalarField& sumOut,
-    scalarField& sumIn
-) const
-{
-    sumOut = 0.0;
-    sumIn = 0.0;
-
-    const labelUList& owner = mesh_.owner();
-    const labelUList& neighbour = mesh_.neighbour();
-
-    forAll(phi, faceI)
-    {
-        const scalar faceFlux = lambda[faceI]*phi[faceI];
-
-        if (faceFlux > 0.0)
-        {
-            sumOut[owner[faceI]] += faceFlux;
-            sumIn[neighbour[faceI]] += faceFlux;
-        }
-        else
-        {
-            sumIn[owner[faceI]] -= faceFlux;
-            sumOut[neighbour[faceI]] -= faceFlux;
-        }
-    }
-
-    forAll(phi.boundaryField(), patchI)
-    {
-        const fvsPatchScalarField& phiP = phi.boundaryField()[patchI];
-        const fvsPatchScalarField& lambdaP = lambda.boundaryField()[patchI];
-        const labelUList& faceCells = mesh_.boundary()[patchI].faceCells();
-
-        forAll(phiP, i)
-        {
-            const scalar faceFlux = lambdaP[i]*phiP[i];
-
-            if (faceFlux > 0.0)
-            {
-                sumOut[faceCells[i]] += faceFlux;
-            }
-            else
-            {
-                sumIn[faceCells[i]] -= faceFlux;
-            }
-        }
-    }
-}
-
-void volPyrolysis::solidFluxBudgets
-(
-    PtrList<surfaceScalarField>& phiYm,
-    PtrList<volScalarField>& RRsolid,
-    tmp<surfaceScalarField>& tPhiYmTotal,
-    tmp<surfaceScalarField>& tPhiSolidVol,
-    tmp<volScalarField>& tAlphaS,
-    tmp<volScalarField>& tRRpor
-)
-{
-    const dimensionedScalar rhoSolidFloor
-    (
-        "rhoSolidFloor",
-        dimDensity,
-        SMALL
-    );
-
-    phiYm.setSize(Ym_.size());
-    RRsolid.setSize(Ym_.size());
-
-    forAll(Ym_, i)
-    {
-        // A fixedYm patch value is Yi*rho*(1 - porosityF) read from the
-        // cell behind it, and the porosity it reads was recovered at the
-        // end of the previous step, so the patch is stale until corrected.
-        // Limiting a stale zero leaves that face unlimited while the
-        // corrected value asks for solid the cell does not have.
-        Ym_[i].correctBoundaryConditions();
-
-        phiYm.set(i, fvc::flux(phiSolid_, Ym_[i], "div(phiSolid)").ptr());
-        RRsolid.set(i, solidChemistry_->RRs(i).ptr());
-    }
-
-    volScalarField totalYm(Ym_[0]);
-
-    tPhiYmTotal = tmp<surfaceScalarField>(new surfaceScalarField(phiYm[0]));
-    surfaceScalarField& phiYmTotal = tPhiYmTotal.ref();
-
-    for (label i = 1; i < Ym_.size(); ++i)
-    {
-        totalYm += Ym_[i];
-        phiYmTotal += phiYm[i];
-    }
-
-    // Read from the mass, not from porosity_: the mass is what the limiter
-    // moves, and recoverPorosity() writes porosity_ from this same
-    // expression - through a "< 1e-4 -> 0" clip that is a second writer.
-    tAlphaS = totalYm/max(rho_, rhoSolidFloor);
-
-    // Chemistry fills and empties cells too. RRpor = -sum_i RRs_i/rho_i is
-    // d(porosity)/dt, so -RRpor is d(alphaS)/dt and a cell that chemistry
-    // is densifying has that much less room for what the flux brings.
-    tRRpor = solidChemistry_->RRpor(T_);
-
-    // The volume an arriving mass occupies is set by where it came from, so
-    // the specific volume is taken upwind of the flux: exact where the
-    // species share a density, second order in the composition step.
-    tPhiSolidVol =
-        phiYmTotal
-       *upwind<scalar>(mesh_, phiSolid_).interpolate
-        (
-            1.0/max(rho_, rhoSolidFloor)
-        );
-}
-
-void volPyrolysis::solidDonorLimit
-(
-    const PtrList<surfaceScalarField>& phiYm,
-    const PtrList<volScalarField>& RRsolid,
-    const surfaceScalarField& lambda,
-    const bool credit,
-    scalarField& lambdaDonor
-) const
-{
-    const scalarField& V = mesh_.V();
-    const scalar rDeltaT = 1.0/time_.deltaTValue();
-
-    scalarField sumOut(mesh_.nCells(), Zero);
-    scalarField sumIn(mesh_.nCells(), Zero);
-
-    lambdaDonor = 1.0;
-
-    forAll(Ym_, i)
-    {
-        accumulateFaceFlux(phiYm[i], lambda, sumOut, sumIn);
-
-        forAll(lambdaDonor, cellI)
-        {
-            // Mass of specie i the cell can part with over this step:
-            // what it holds, less what chemistry takes from it.
-            const scalar canLeave = max
-            (
-                0.0,
-                (Ym_[i][cellI]*rDeltaT + RRsolid[i][cellI])*V[cellI]
-              + (credit ? sumIn[cellI] : 0.0)
-            );
-
-            // Nothing leaving means nothing to scale. A factor of zero
-            // here would report a limit on faces carrying no solid at all.
-            if (sumOut[cellI] > SMALL)
-            {
-                lambdaDonor[cellI] = min
-                (
-                    lambdaDonor[cellI],
-                    min(1.0, canLeave/sumOut[cellI])
-                );
-            }
-        }
-    }
-}
-
-void volPyrolysis::solidReceiverLimit
-(
-    const surfaceScalarField& phiSolidVol,
-    const volScalarField& alphaS,
-    const volScalarField& RRpor,
-    const surfaceScalarField& lambda,
-    const bool credit,
-    scalarField& lambdaReceiver
-) const
-{
-    const scalarField& V = mesh_.V();
-    const scalar rDeltaT = 1.0/time_.deltaTValue();
-    const scalar alphaSMax = 1.0 - minPorosity_;
-
-    scalarField sumOut(mesh_.nCells(), Zero);
-    scalarField sumIn(mesh_.nCells(), Zero);
-
-    accumulateFaceFlux(phiSolidVol, lambda, sumOut, sumIn);
-
-    forAll(lambdaReceiver, cellI)
-    {
-        // Solid volume the cell still has room for over this step, after
-        // chemistry has taken its share of it.
-        const scalar room = max
-        (
-            0.0,
-            ((alphaSMax - alphaS[cellI])*rDeltaT + RRpor[cellI])*V[cellI]
-          + (credit ? sumOut[cellI] : 0.0)
-        );
-
-        lambdaReceiver[cellI] =
-            sumIn[cellI] > SMALL
-          ? min(1.0, room/sumIn[cellI])
-          : 1.0;
-    }
-}
-
-void volPyrolysis::applySolidFaceLimit
-(
-    const surfaceScalarField& phiYmTotal,
-    const scalarField& lambdaDonor,
-    const scalarField& lambdaReceiver,
-    scalarField& allLambda,
-    surfaceScalarField& lambda
-) const
-{
-    const labelUList& owner = mesh_.owner();
-    const labelUList& neighbour = mesh_.neighbour();
-
-    scalarField& lambdaIn = lambda;
-    surfaceScalarField::Boundary& lambdaBf = lambda.boundaryFieldRef();
-
-    forAll(lambdaIn, faceI)
-    {
-        const label own = owner[faceI];
-        const label nei = neighbour[faceI];
-
-        if (phiYmTotal[faceI] > 0.0)
-        {
-            lambdaIn[faceI] = min
-            (
-                lambdaIn[faceI],
-                min(lambdaDonor[own], lambdaReceiver[nei])
-            );
-        }
-        else
-        {
-            lambdaIn[faceI] = min
-            (
-                lambdaIn[faceI],
-                min(lambdaDonor[nei], lambdaReceiver[own])
-            );
-        }
-    }
-
-    forAll(lambdaBf, patchI)
-    {
-        fvsPatchScalarField& lambdaP = lambdaBf[patchI];
-        const fvsPatchScalarField& phiP =
-            phiYmTotal.boundaryField()[patchI];
-        const labelUList& faceCells =
-            mesh_.boundary()[patchI].faceCells();
-
-        forAll(lambdaP, i)
-        {
-            // Only this side of the patch is reachable here. A coupled
-            // patch takes the other side's factor from the sync below; a
-            // real boundary has no other side, so an inlet is held by the
-            // receiving cell's room and an outlet by what the draining
-            // cell holds.
-            lambdaP[i] = min
-            (
-                lambdaP[i],
-                phiP[i] > 0.0
-              ? lambdaDonor[faceCells[i]]
-              : lambdaReceiver[faceCells[i]]
-            );
-        }
-    }
-
-    // lambda slices allLambda, patch faces included, so the two sides of a
-    // coupled face meet here and both keep the tighter factor.
-    syncTools::syncFaceList(mesh_, allLambda, minEqOp<scalar>());
-}
-
-void volPyrolysis::reportSolidFluxLimiter
-(
-    const PtrList<surfaceScalarField>& phiYm,
-    const PtrList<volScalarField>& RRsolid,
-    const surfaceScalarField& phiSolidVol,
-    const volScalarField& alphaS,
-    const volScalarField& RRpor,
-    const surfaceScalarField& lambda
-) const
-{
-    const scalarField& V = mesh_.V();
-    const scalar deltaT = time_.deltaTValue();
-    const scalar alphaSMax = 1.0 - minPorosity_;
-
-    const scalarField& lambdaIn = lambda;
-    const surfaceScalarField::Boundary& lambdaBf = lambda.boundaryField();
-
-    scalar nLimited = 0.0;
-    scalar withheld = 0.0;
-    scalar minLambda = 1.0;
-
-    forAll(lambdaIn, faceI)
-    {
-        minLambda = min(minLambda, lambdaIn[faceI]);
-
-        if (lambdaIn[faceI] < 1.0 - SMALL)
-        {
-            nLimited += 1.0;
-            withheld +=
-                (1.0 - lambdaIn[faceI])*mag(phiSolidVol[faceI])*deltaT;
-        }
-    }
-
-    forAll(lambdaBf, patchI)
-    {
-        // A coupled face is one face seen from two sides, so each side
-        // carries half of it and the totals come out per physical face.
-        const scalar weight =
-            mesh_.boundary()[patchI].coupled() ? 0.5 : 1.0;
-
-        const fvsPatchScalarField& lambdaP = lambdaBf[patchI];
-        const fvsPatchScalarField& phiVolP =
-            phiSolidVol.boundaryField()[patchI];
-
-        forAll(lambdaP, i)
-        {
-            minLambda = min(minLambda, lambdaP[i]);
-
-            if (lambdaP[i] < 1.0 - SMALL)
-            {
-                nLimited += weight;
-                withheld +=
-                    weight*(1.0 - lambdaP[i])*mag(phiVolP[i])*deltaT;
-            }
-        }
-    }
-
-    // What the sweeps did not reach, read off the state the limited flux
-    // will produce rather than off the inequality the factors were built
-    // from. The sweep count is capped, so this is the honest answer.
-    scalar maxUndershoot = 0.0;
-    scalar YmScale = SMALL;
-
-    scalarField sumOut(mesh_.nCells(), Zero);
-    scalarField sumIn(mesh_.nCells(), Zero);
-
-    forAll(Ym_, i)
-    {
-        YmScale = max(YmScale, gMax(Ym_[i]));
-
-        accumulateFaceFlux(phiYm[i], lambda, sumOut, sumIn);
-
-        forAll(Ym_[i], cellI)
-        {
-            const scalar YmNew =
-                Ym_[i][cellI]
-              + deltaT
-               *(
-                    RRsolid[i][cellI]
-                  + (sumIn[cellI] - sumOut[cellI])/V[cellI]
-                );
-
-            maxUndershoot = max(maxUndershoot, -YmNew);
-        }
-    }
-
-    accumulateFaceFlux(phiSolidVol, lambda, sumOut, sumIn);
-
-    scalar maxOvershoot = 0.0;
-
-    forAll(alphaS, cellI)
-    {
-        const scalar alphaSNew =
-            alphaS[cellI]
-          + deltaT
-           *(
-              - RRpor[cellI]
-              + (sumIn[cellI] - sumOut[cellI])/V[cellI]
-            );
-
-        maxOvershoot = max(maxOvershoot, alphaSNew - alphaSMax);
-    }
-
-    nLimited = returnReduce(nLimited, sumOp<scalar>());
-    withheld = returnReduce(withheld, sumOp<scalar>());
-    minLambda = returnReduce(minLambda, minOp<scalar>());
-    maxOvershoot = returnReduce(maxOvershoot, maxOp<scalar>());
-    maxUndershoot = returnReduce(maxUndershoot, maxOp<scalar>());
-
-    if (nLimited > 0.5)
-    {
-        Info<< "solid flux limiter: faces limited = "
-            << label(nLimited + 0.5)
-            << ", min scale factor = " << minLambda
-            << ", solid volume withheld = " << withheld << " m3";
-
-        if
-        (
-            maxOvershoot > solidStateTolerance_
-         || maxUndershoot > solidStateTolerance_*YmScale
-        )
-        {
-            Info<< ", NOT converged in " << nSolidFluxLimiterCorrectors_
-                << " sweeps: residual packing = " << maxOvershoot
-                << ", residual mass deficit = " << maxUndershoot;
-        }
-
-        Info<< endl;
-    }
-}
-
 void volPyrolysis::limitSolidVolFlux()
 {
-    phiSolid_ = solidVolFlux();
-
-    if (!active_ || !advectSolidFields_ || nSolidFluxLimiterCorrectors_ < 1)
-    {
-        return;
-    }
-
-    // The budgets are written in the terms of the equation they limit,
-    // Ym_i^new = Ym_i + dt*(RRs_i - div(phiYm_i)), so phiYm carries the
-    // transport scheme's own face fluxes. Scaling phiSolid_ scales every
-    // one of them by the same factor: the scheme reads the flux only for
-    // the upwind direction, which a non-negative scale leaves alone.
-    PtrList<surfaceScalarField> phiYm;
-    PtrList<volScalarField> RRsolid;
-    tmp<surfaceScalarField> tPhiYmTotal;
-    tmp<surfaceScalarField> tPhiSolidVol;
-    tmp<volScalarField> tAlphaS;
-    tmp<volScalarField> tRRpor;
-
-    solidFluxBudgets
-    (
-        phiYm,
-        RRsolid,
-        tPhiYmTotal,
-        tPhiSolidVol,
-        tAlphaS,
-        tRRpor
-    );
-
-    const surfaceScalarField& phiYmTotal = tPhiYmTotal();
-    const surfaceScalarField& phiSolidVol = tPhiSolidVol();
-    const volScalarField& alphaS = tAlphaS();
-    const volScalarField& RRpor = tRRpor();
-
-    scalarField allLambda(mesh_.nFaces(), 1.0);
-
-    slicedSurfaceScalarField lambda
-    (
-        IOobject
-        (
-            "solidFluxLimiter",
-            time_.timeName(),
-            mesh_,
-            IOobject::NO_READ,
-            IOobject::NO_WRITE,
-            false
-        ),
-        mesh_,
-        dimless,
-        allLambda,
-        false               // slice the couples, so syncFaceList sees them
-    );
-
-    scalarField lambdaDonor(mesh_.nCells(), 1.0);
-    scalarField lambdaReceiver(mesh_.nCells(), 1.0);
-
-    for (label sweep = 0; sweep < nSolidFluxLimiterCorrectors_; ++sweep)
-    {
-        // All but the last sweep credit a cell with the flux crossing its
-        // faces the other way - what arrives may also leave, what leaves
-        // makes room - which lets a jam travel back up the bed inside one
-        // step. The last sweep drops the credit, and its bound is the one
-        // that holds for the flux finally applied.
-        const bool credit = (sweep < nSolidFluxLimiterCorrectors_ - 1);
-
-        solidDonorLimit(phiYm, RRsolid, lambda, credit, lambdaDonor);
-
-        solidReceiverLimit
-        (
-            phiSolidVol,
-            alphaS,
-            RRpor,
-            lambda,
-            credit,
-            lambdaReceiver
-        );
-
-        applySolidFaceLimit
-        (
-            phiYmTotal,
-            lambdaDonor,
-            lambdaReceiver,
-            allLambda,
-            lambda
-        );
-    }
-
-    reportSolidFluxLimiter
-    (
-        phiYm,
-        RRsolid,
-        phiSolidVol,
-        alphaS,
-        RRpor,
-        lambda
-    );
-
-    phiSolid_ *= lambda;
-}
-
-label volPyrolysis::firstInvalidCell
-(
-    const volScalarField& fld,
-    const scalar lower,
-    const scalar upper,
-    const scalarField* mask
-) const
-{
-    forAll(fld, cellI)
-    {
-        if (mask && (*mask)[cellI] <= 0.0)
-        {
-            continue;
-        }
-
-        const scalar value = fld[cellI];
-
-        if (!std::isfinite(value) || value < lower || value > upper)
-        {
-            return cellI;
-        }
-    }
-
-    return -1;
-}
-
-void volPyrolysis::reportInvalidSolidState
-(
-    const word& stage,
-    const volScalarField& fld,
-    const label cellI,
-    const string& context
-) const
-{
-    FatalErrorInFunction
-        << "Impossible solid state produced by " << stage << nl << nl
-        << "    time         = " << time_.timeName() << nl
-        << "    deltaT       = " << time_.deltaTValue() << nl
-        << "    field        = " << fld.name() << nl
-        << "    value        = " << fld[cellI] << nl
-        << "    processor    = " << Pstream::myProcNo() << nl
-        << "    cell         = " << cellI << nl
-        << "    cell centre  = " << mesh_.C()[cellI] << nl
-        << "    cell volume  = " << mesh_.V()[cellI] << nl
-        << context.c_str() << nl
-        << "The solid transport equations are explicit and unbounded. Either"
-        << " the solid velocity Us is invalid, or the explicit update has"
-        << " overshot at this time step." << nl
-        << exit(FatalError);
+    phiSolid_ = solidFluxLimiter_->limit();
 }
 
 void volPyrolysis::recoverPorosity()
@@ -761,7 +187,7 @@ void volPyrolysis::recoverPorosity()
             // Before the "< 1e-4 -> 0" clip below, which would absorb an
             // undershoot without trace. por > 1 is negative solid mass;
             // por < 0 is a cell packed past solid by the transport.
-            const label badCell = firstInvalidCell
+            const label badCell = firstInvalidSolidCell
             (
                 por,
                 -solidStateTolerance_,
@@ -781,6 +207,8 @@ void volPyrolysis::recoverPorosity()
 
                 reportInvalidSolidState
                 (
+                    mesh_,
+                    time_,
                     "the porosity recovery in recoverPorosity()",
                     por,
                     badCell,
@@ -1572,7 +1000,7 @@ void volPyrolysis::solveSpeciesMass()
                 // the amount of solid actually present in the field.
                 const scalar YmScale = max(gMax(Ym_i), SMALL);
 
-                const label badCell = firstInvalidCell
+                const label badCell = firstInvalidSolidCell
                 (
                     Ym_i,
                     -solidStateTolerance_*YmScale,
@@ -1591,6 +1019,8 @@ void volPyrolysis::solveSpeciesMass()
 
                     reportInvalidSolidState
                     (
+                        mesh_,
+                        time_,
                         "the solid specie mass equation in solveSpeciesMass()",
                         Ym_i,
                         badCell,
@@ -1719,7 +1149,7 @@ void volPyrolysis::preSolveEnergy()
                 // tail of a bed the capacity decays towards zero while the
                 // solver's error does not, so Ts comes back a few 1e-6 K
                 // negative. Judged against the scale the guard declares.
-                const label badCell = firstInvalidCell
+                const label badCell = firstInvalidSolidCell
                 (
                     T_,
                    -solidStateTolerance_*maxSolidTemperature_,
@@ -1740,6 +1170,8 @@ void volPyrolysis::preSolveEnergy()
 
                     reportInvalidSolidState
                     (
+                        mesh_,
+                        time_,
                         "Ts = solidH/rhoCp at the head of preSolveEnergy()",
                         T_,
                         badCell,
@@ -1811,7 +1243,7 @@ void volPyrolysis::preSolveEnergy()
 
             if (failOnInvalidSolidState_)
             {
-                const label badCell = firstInvalidCell
+                const label badCell = firstInvalidSolidCell
                 (
                     T_,
                    -solidStateTolerance_*maxSolidTemperature_,
@@ -1835,6 +1267,8 @@ void volPyrolysis::preSolveEnergy()
 
                     reportInvalidSolidState
                     (
+                        mesh_,
+                        time_,
                         "the solid energy equation in preSolveEnergy()",
                         T_,
                         badCell,
@@ -1897,7 +1331,7 @@ void volPyrolysis::preSolveEnergy()
                 const volScalarField& sH = solidH_();
                 const scalar sHScale = max(gMax(sH), SMALL);
 
-                const label badCell = firstInvalidCell
+                const label badCell = firstInvalidSolidCell
                 (
                     sH,
                     -solidStateTolerance_*sHScale,
@@ -1916,6 +1350,8 @@ void volPyrolysis::preSolveEnergy()
 
                     reportInvalidSolidState
                     (
+                        mesh_,
+                        time_,
                         "the solid enthalpy advection in preSolveEnergy()",
                         sH,
                         badCell,
@@ -1988,7 +1424,7 @@ void volPyrolysis::postSolveEnergy()
                 // A cell holding no solid holds no solid enthalpy, so the
                 // lower bound has to admit zero here.
                 const label badCell =
-                    firstInvalidCell(T_, 0.0, maxSolidTemperature_);
+                    firstInvalidSolidCell(T_, 0.0, maxSolidTemperature_);
 
                 if (badCell != -1)
                 {
@@ -2003,6 +1439,8 @@ void volPyrolysis::postSolveEnergy()
 
                     reportInvalidSolidState
                     (
+                        mesh_,
+                        time_,
                         "the Ts recovery in postSolveEnergy()",
                         T_,
                         badCell,
@@ -2138,7 +1576,6 @@ volPyrolysis::volPyrolysis
     replenishSwitch_(false),
     collapseMovesSolidMass_(false),
     emptyFlippedCells_(false),
-    advectSolidFields_(true),
     gasTemperatureBelowCriticalPorosity_(false),
     failOnInvalidSolidState_(true),
     solidStateTolerance_(1e-8),
@@ -2328,6 +1765,7 @@ volPyrolysis::volPyrolysis
         mesh_,
         dimensionedScalar("zero", dimVolume/dimTime, 0.0)
     ),
+    solidFluxLimiter_(nullptr),
     demActive_(false),
     lambdaDotPtr_(nullptr),
     lostSolidMass_(dimensionedScalar("zero", dimMass, 0.0)),
@@ -2355,7 +1793,6 @@ volPyrolysis::volPyrolysis
         coeffs().lookupOrDefault("collapseMovesSolidMass",false);
     emptyFlippedCells_ =
         coeffs().lookupOrDefault("emptyFlippedCells",false);
-    advectSolidFields_ = coeffs().lookupOrDefault("advectSolidFields",true);
     gasTemperatureBelowCriticalPorosity_ =
         coeffs().lookupOrDefault("gasTemperatureBelowCriticalPorosity",false);
     failOnInvalidSolidState_ =
@@ -2364,9 +1801,6 @@ volPyrolysis::volPyrolysis
         coeffs().lookupOrDefault<scalar>("solidStateTolerance",1e-8);
     maxSolidTemperature_ =
         coeffs().lookupOrDefault<scalar>("maxSolidTemperature",1e5);
-    nSolidFluxLimiterCorrectors_ =
-        coeffs().lookupOrDefault<label>("nSolidFluxLimiterCorrectors",3);
-    minPorosity_ = coeffs().lookupOrDefault<scalar>("minPorosity",0.0);
     critPorosity_ = coeffs().lookupOrDefault("criticalPorosity",0.9999);
     poroProtectSolidInflowFluxTolerance_ =
         coeffs().lookupOrDefault
@@ -2388,15 +1822,31 @@ volPyrolysis::volPyrolysis
     Info << "replenish                " << replenishSwitch_    << endl;
     Info << "collapseMovesSolidMass   " << collapseMovesSolidMass_ << endl;
     Info << "emptyFlippedCells        " << emptyFlippedCells_  << endl;
-    Info << "advectSolidFields        " << advectSolidFields_  << endl;
     Info << "gasTemperatureBelowCriticalPorosity  "
          << gasTemperatureBelowCriticalPorosity_ << endl;
     Info << "failOnInvalidSolidState  " << failOnInvalidSolidState_ << endl;
     Info << "solidStateTolerance      " << solidStateTolerance_ << endl;
     Info << "maxSolidTemperature      " << maxSolidTemperature_ << endl;
-    Info << "nSolidFluxLimiterCorrectors " << nSolidFluxLimiterCorrectors_
-         << endl;
-    Info << "minPorosity              " << minPorosity_ << endl;
+
+    // Reads its own keys from the same dict and logs them, so the settings
+    // it owns are reported where the rest of them are. Ym_ is filled below;
+    // the limiter binds the list, not its entries, so it does not read them
+    // until limit() is first called.
+    solidFluxLimiter_.reset
+    (
+        new SolidFluxLimiter
+        (
+            coeffs(),
+            mesh_,
+            Us_,
+            Ym_,
+            rho_,
+            T_,
+            solidChemistry_(),
+            active_
+        )
+    );
+
     Info << endl;
 
     forAll(Ys_, fieldI)
